[EDDI] Fixed preprocess
Insert in toHardenVariables in a less aggressive way, only if the variable can be considered an "alias" of the variables annotated with `to_harden`. In our case we consider it to be an alias if it is the result of a load, the other operator of a store or the result of the GetElementPointer instruction.
Also, add functions and variables in their toHarden sets only if isToDuplicate returns false, meaning that they have really to be hardened and not just duplicated their calls